@@ -8,7 +8,7 @@ class Settings(BaseSettings):
     # Spotify API Configuration
     spotify_client_id: str
     spotify_client_secret: str
-    spotify_redirect_uri: str = "http://localhost:8000/auth/callback"
+    spotify_redirect_uri: str = "http://127.0.0.1:8000/auth/callback"
 
     # Application Security
     secret_key: str
